requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
google-cloud-secret-manager
google-cloud-resource-manager
google-cloud-firestore
cryptography
uuid